Implement support for as/1 dynamic bindings
This is a useful tool for query composition added in Ecto 3.4. This is backwards compatible with the
older Ecto version, because "aliases" is an already existing field on the Ecto.Query struct.

The `as` binding passes its value as a literal instead of as a TaggedValue, so parse_value is expanded
to support raw values.